fix npm build error
